+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Yii Validator Change Log
 
-## 2.6.2 under development
+## 2.7.0 under development
 
-- no changes in this release.
+- New #824: Add `ValidationContext::validateInCurrentScope()` method to validate a value in the current scope (@vjik)
+- Bug #824: Validate rules inside `StopOnError`, `Composite` and `AnyRule` in the current scope, same as without these
+  wrappers: keep the current data set and property in validation context, don't iterate over public properties of
+  an object value and don't call its `PostValidationHookInterface::processValidationResult()` (@vjik)
 
 ## 2.6.1 September 22, 2026